Feat: Logger 추가
* Console 함수 호출 영역 Logger로 대체

+++ src/Cleaner.Util/DirectoryCleaner.cs
@@ -5,15 +5,18 @@ namespace Cleaner.Util
     {
         private double _desiredFreeRatio;
         private string _managedDirectoryPath;
-        private SearchOption _searchOption = SearchOption.TopDirectoryOnly;
+        private SearchOption _searchOption = SearchOption.AllDirectories;
         private System.Timers.Timer _timer;
 
         public long TotalSpace { get; private set; }
         public long AvailableSpace { get; private set; }
 
-        private void CheckDriveStatus(string path)
+        public delegate void TaskCompleteHandler();
+        public TaskCompleteHandler TaskCompleteEvent;
+
+        public void CheckDriveStatus()
         {
-            DriveInfo drive = new DriveInfo(path);
+            DriveInfo drive = new DriveInfo(Path.GetPathRoot(_managedDirectoryPath));
             AvailableSpace = drive.AvailableFreeSpace;
             TotalSpace = drive.TotalSize;
         }
@@ -21,7 +24,7 @@ namespace Cleaner.Util
         public void SetManagedDirectoryPath(string directoryPath)
         {
             _managedDirectoryPath = directoryPath;
-            CheckDriveStatus(Path.GetPathRoot(directoryPath));
+            CheckDriveStatus();
         }
 
         public void SetDesiredFreeRatio(double desiredFreeRatio)
@@ -29,13 +32,15 @@ namespace Cleaner.Util
             _desiredFreeRatio = desiredFreeRatio;
         }
 
-        public void Run()
+        public async void Run()
         {
-            Cleanup();
+            Logger.Instance.Print(Logger.LogLevel.INFO, $"작업 시작");
+            await Task.Run(Cleanup);
         }
 
         public void Run(double time)
         {
+            Logger.Instance.Print(Logger.LogLevel.INFO, $"타이머 작업 시작");
             _timer = new System.Timers.Timer(time);
             _timer.Elapsed += (sender, e) => Cleanup();
             _timer.Start();
@@ -43,6 +48,7 @@ namespace Cleaner.Util
 
         public void Stop()
         {
+            Logger.Instance.Print(Logger.LogLevel.INFO, $"타이머 작업 종료");
             _timer.Elapsed -= (sender, e) => Cleanup();
             _timer.Stop();
         }
@@ -51,35 +57,43 @@ namespace Cleaner.Util
         {
             try
             {
+                long bytesToFree = ComputeRequiredFreeSpace();
+                long freedBytes = 0;
+
+                if (bytesToFree <= 0)
+                {
+                    return;
+                }
+
                 string[] files = Directory.GetFiles(_managedDirectoryPath, "*.*", _searchOption);
                 var filesInfo = files.Select(file => new FileInfo(file))
                                      .OrderBy(f => f.LastWriteTime)
                                      .ToList();
 
-                long bytesToFree = ComputeRequiredFreeSpace();
-                long freedBytes = 0;
-
+                int deletedFileCount = 0;
                 foreach (var fileInfo in filesInfo)
                 {
                     freedBytes += fileInfo.Length;
-                    Console.WriteLine($"Deleting: {fileInfo.FullName}, Size: {fileInfo.Length / 1024} KB");
                     fileInfo.Delete();
-
+                    deletedFileCount++;
                     if (freedBytes >= bytesToFree)
                     {
                         break; // 필요한 용량 확보 시 삭제 중지
                     }
                 }
+
+                Logger.Instance.Print(Logger.LogLevel.INFO, $"작업 완료. {deletedFileCount} 개 파일 삭제");
+                TaskCompleteEvent?.Invoke();
             }
             catch(Exception ex)
             {
-                Console.WriteLine("An error occurred during cleanup: " + ex.Message);
+                Logger.Instance.Print(Logger.LogLevel.ERROR, "An error occurred during cleanup: " + ex.Message);
             }
         }
 
         private long ComputeRequiredFreeSpace()
         {
-            DriveInfo drive = new DriveInfo("C");
+            DriveInfo drive = new DriveInfo(Path.GetPathRoot(_managedDirectoryPath));
             return (long)((1 - _desiredFreeRatio) * TotalSpace) - drive.AvailableFreeSpace;
         }
     }
